refactor: hardware basico

+++ src/br/inatel/pcmania/formatadores/Formatadores.java
@@ -1,7 +1,7 @@
 package br.inatel.pcmania.formatadores;
 
 public class Formatadores{
-  public static String formatarMemoria(int capacidadeGB){
+  public static String formatarMemoria(float capacidadeGB){
     if(capacidadeGB > 1000){
       return capacidadeGB/1000 + " TB";
     }
